Move externalDirectories to esbuild-options
