+++ src/esbuild-options.ts
@@ -51,4 +51,27 @@ export const findEntryPoints = (root: string): Record<string, string> => {
   return result;
 }
 
+// TODO: feels like this really should be passed a root too, to become the cwd for globSync
+export const externalDirectories = (): string[] => {
+  const assetDirsPattern = [
+    path.join("app", "assets", "*"),
+    path.join("slices", "*", "assets", "*"),
+  ]
+
+  const excludeDirs = ['js', 'css'];
+
+  try {
+    const dirs = globSync(assetDirsPattern, { nodir: false });
+    const filteredDirs = dirs.filter((dir) => {
+      const dirName = dir.split(path.sep).pop();
+      return !excludeDirs.includes(dirName!);
+    });
+
+    return filteredDirs.map((dir) => path.join(dir, "*"));
+  } catch (err) {
+    console.error('Error listing external directories:', err);
+    return [];
+  }
+};
+
